[Windows] Kent suggested casting the variables rather than changing the format specificier

+++ YapDatabase/Utilities/YapBidirectionalCache.m
@@ -462,8 +462,8 @@ const YapBidirectionalCacheCallBacks kYapBidirectionalCacheDefaultCallBacks = (Y
 			if (leastRecentCacheItem == nil)
 				leastRecentCacheItem = newItem;
 			
-			YDBLogVerbose(@"key(%@) <- new mostRecent [%lld of %lu]",
-			              key, CFDictionaryGetCount(key_obj_dict), (unsigned long)countLimit);
+			YDBLogVerbose(@"key(%@) <- new mostRecent [%ld of %lu]",
+			              key, (long)CFDictionaryGetCount(key_obj_dict), (unsigned long)countLimit);
 		}
 	}
